# TSDB: make the global hash lookup table smaller