--- a/tsdb/head.go
+++ b/tsdb/head.go
@@ -1736,6 +1736,9 @@ func (m seriesHashmap) set(hash uint64, s *memSeries) {
 		m.unique[hash] = s
 		return
 	}
+	if m.conflicts == nil {
+		m.conflicts = make(map[uint64][]*memSeries)
+	}
 	l := m.conflicts[hash]
 	for i, prev := range l {
 		if labels.Equal(prev.lset, s.lset) {
@@ -1815,7 +1818,7 @@ func newStripeSeries(stripeSize int, seriesCallback SeriesLifecycleCallback) *st
 	for i := range s.hashes {
 		s.hashes[i] = seriesHashmap{
 			unique:    map[uint64]*memSeries{},
-			conflicts: map[uint64][]*memSeries{},
+			conflicts: nil, // Initialized on demand in set().
 		}
 	}
 	return s